requirements.txt
Django==2.0.2
python-dotenv==0.7.1
dj-database-url==0.4.2
django-crispy-forms==1.7.0
djangorestframework==3.7.7
django-filter==1.1.0




